Implementa controle de saque em PaymentAPI (#7)
* rework on WithdrawManager and test wip

* getWithdrawsByIdAsync wip

* create AddAccountId migration

* update in WithdrawManager wip

* new controller for Withdraws

* update in AccountController test and his manager

* implemented WithdrawManager test

* ShouldListById test

* implemented tests for WithdrawManager and fix

* movement wip

* extract wip (withdraw broken)

* solved problem in withdraw and add Extracts

* solved problems on old tests

* extract tests wip

+++ api/Models/Installment/Installment.cs
@@ -12,10 +12,10 @@ public class Installment
     public int Id { get; set; }  // Id correspondente à transação.
 
     [Required]
-    public float InstallmentGrossValue { get; set; } // Valor bruto da parcela.
+    public decimal InstallmentGrossValue { get; set; } // Valor bruto da parcela.
 
     [Required]
-    public float InstallmentNetValue { get; set; } // Valor líquido da parcela.
+    public decimal InstallmentNetValue { get; set; } // Valor líquido da parcela.
 
     [Required]
     public int InstallmentNumber { get; set; } // Número da parcela
